Branch1

+++ Frontend/js/edicion.js
@@ -1,5 +1,18 @@
 import { ResumenFinanciero, FlujoDeEfectivo, EstadoDeResultados, Ingresos, CostosDirectos, GastosAdministrativos, Recursos } from "../js/clases.js";
 
+export function Calcular(a){
+  //console.log("calculando valores");
+  recursos.calcularValorRecursos();
+  recursos.calcularValorCostos();
+  recursos.calcularValorResumen();
+
+  gastos.calcularValor();
+  costos.calcularValor();
+  ingresos.calcularValor();
+  estado.calcularValor();
+  flujo.calcularValor();
+  resumen.calcularValor();
+}
 /*
 Instanciacion de los objetos definidos en clases.js
 pertenecientes a cada seccion del ticket
@@ -32,6 +45,7 @@ ingresos.crearIngresos();
 costos.crearCostosDirectos();
 gastos.crearGastosAdministrativos();
 recursos.crearRecursos();
+Calcular(); 
 
 //Guardado de los botones para posteriormente agregarles el event listener
 const guardarPresupuesto = document.getElementById('guardarPresupuesto');
@@ -154,25 +168,29 @@ agregarMes.addEventListener('click', () => {
     rolesRecursos.push(respRecursos);
 
   }
-  
+  Calcular();  
 });
 
 //Eliminacin de columnas
 eliminarMes.addEventListener('click', () => {
-  //Eliminacion de columna en cad parte del DOM
-  flujo.eliminarColumna();
-  estado.eliminarColumna();
-  ingresos.eliminarColumna();
-  costos.eliminarColumna();
-  gastos.eliminarColumna();
-  recursos.eliminarColumna();
 
-  //Eliminacion de columna en cada array correspondiente
-  mesesContemplados.pop();
-  conceptosIngresos.pop();
-  conceptosCostos.pop();
-  conceptosGastos.pop();
-  rolesRecursos.pop();
+  if (window.confirm("¿Esta seguro de que desea eliminar la columna?")) {
+    //Eliminacion de columna en cad parte del DOM
+    flujo.eliminarColumna();
+    estado.eliminarColumna();
+    ingresos.eliminarColumna();
+    costos.eliminarColumna();
+    gastos.eliminarColumna();
+    recursos.eliminarColumna();
+    //Eliminacion de columna en cada array correspondiente
+    mesesContemplados.pop();
+    conceptosIngresos.pop();
+    conceptosCostos.pop();
+    conceptosGastos.pop();
+    rolesRecursos.pop();
+    //Calculo de nuevos valores tras eliminacion de columna
+    Calcular();
+  }  
 });
 
 
@@ -200,27 +218,34 @@ agregarIngreso.addEventListener('click', () => {
         conceptosIngresos[index].push(element);
       }  
     });
-  }  
+  }
+  Calcular();  
 });
 
 //eliminacion de una fila correspondiente a un ingreso
-eliminarIngreso.addEventListener('click', () => {  
-  //Si ya se selecciono un ingreso en la tabla se elimina dicha seleccion
-  if (ingresos.focusActual != null) {  
-    //Eliminacion del elemento Json correspondiente en cada arreglo de Jsons
-    conceptosIngresos.forEach(element => {
-      element.forEach((elemento, index) => {
-        if(elemento.concepto == ingresos.focusActual){
-          element.splice(index,1);
-        }        
-      }); 
-    });
-    //Eliminacion de la fila seleccionada en el DOM
-    ingresos.eliminarFila(ingresos.focusActual);
-  }
-  //si no, se solicita seleccionar un ingreso de la tabla
-  else
-    alert("selecciona un concepto de ingreso primero");  
+eliminarIngreso.addEventListener('click', () => { 
+
+  if (window.confirm("¿Esta seguro de que desea eliminar el ingreso?")) {
+    //Si ya se selecciono un ingreso en la tabla se elimina dicha seleccion
+    if (ingresos.focusActual != null) { 
+      //Eliminacion del elemento Json correspondiente en cada arreglo de Jsons
+      conceptosIngresos.forEach(element => {
+        element.forEach((elemento, index) => {
+          if(elemento.concepto == ingresos.focusActual){
+            element.splice(index,1);
+          }        
+        }); 
+      });
+      //Eliminacion de la fila seleccionada en el DOM
+      ingresos.eliminarFila(ingresos.focusActual);
+    }
+    //si no, se solicita seleccionar un ingreso de la tabla
+    else
+      alert("selecciona un concepto de ingreso primero");  
+
+    //Calculo de valores tras eliminacion del ingreso
+    Calcular();       
+  }  
 });
 
 //creacion de una fila correspondiente a un costo
@@ -228,9 +253,19 @@ agregarCosto.addEventListener('click', () => {
   do{
     var concepto = window.prompt("Por favor ingrese el concepto del costo", "");
   }while(concepto == "");
-  if (concepto != null) {
+
+  if (costos.opcionActual == 3) {
+    do{
+      var porcentaje = window.prompt("Por favor ingrese el porcentaje", "");
+    }while(isNaN(porcentaje) || porcentaje < 1 || porcentaje > 100);    
+  }else{ 
+    var porcentaje = 0;
+  }
+
+
+  if (concepto != null && porcentaje != null) {
     if (costos.opcionActual != null) {
-      let respuesta = costos.agregarFila(concepto, costos.opcionActual, 0);      
+      let respuesta = costos.agregarFila(concepto, costos.opcionActual, 0, porcentaje);      
       respuesta.forEach((element, index) => {
         if (index >= conceptosCostos.length) {
           let array = [];
@@ -243,23 +278,30 @@ agregarCosto.addEventListener('click', () => {
     }else{
       alert("Primero selecciona una opcion");
     }    
-  }  
+  } 
+  Calcular(); 
 });
 
 //eliminacion de una fila correspondiente a un costo
-eliminarCosto.addEventListener('click', () => {  
-  if (costos.focusActual != null) {  
-    conceptosCostos.forEach(element => {
-      element.forEach((elemento, index) => {
-        if(elemento.concepto == costos.focusActual){
-          element.splice(index,1);
-        }        
-      }); 
-    });
-    costos.eliminarFila(costos.focusActual);
-  }
-  else
-    alert("selecciona un concepto de costo primero");  
+eliminarCosto.addEventListener('click', () => {
+
+  if (window.confirm("¿Esta seguro de que desea eliminar el costo?")) {
+    if (costos.focusActual != null) {  
+      conceptosCostos.forEach(element => {
+        element.forEach((elemento, index) => {
+          if(elemento.concepto == costos.focusActual){
+            element.splice(index,1);
+          }        
+        }); 
+      });
+      costos.eliminarFila(costos.focusActual);
+    }
+    else
+      alert("selecciona un concepto de costo primero");  
+    
+    //Calculo de valores tras eliminacion de costo
+    Calcular();    
+  }  
 });
 
 //creacion de una fila correspondiente a un gasto
@@ -292,23 +334,30 @@ agregarGasto.addEventListener('click', () => {
     }else{
       alert("Primero selecciona una opcion");
     }    
-  }  
+  }
+  Calcular();   
 });
 
 //eliminacion de una fila correspondiente a un gasto
-eliminarGasto.addEventListener('click', () => {  
-  if (gastos.focusActual != null) {  
-    conceptosGastos.forEach(element => {
-      element.forEach((elemento, index) => {
-        if(elemento.concepto == gastos.focusActual){
-          element.splice(index,1);
-        }
-      }); 
-    });
-    gastos.eliminarFila(gastos.focusActual);
-  }
-  else
-    alert("selecciona un concepto de gasto primero");  
+eliminarGasto.addEventListener('click', () => {
+
+  if (window.confirm("¿Esta seguro de que desea eliminar el gasto?")) {
+    if (gastos.focusActual != null) {  
+      conceptosGastos.forEach(element => {
+        element.forEach((elemento, index) => {
+          if(elemento.concepto == gastos.focusActual){
+            element.splice(index,1);
+          }
+        }); 
+      });
+      gastos.eliminarFila(gastos.focusActual);
+    }
+    else
+      alert("selecciona un concepto de gasto primero");
+    
+    //Calculo de valores tras eliminacion de gasto  
+    Calcular();    
+  }  
 });
 
 //creacion de una fila correspondiente a un recurso
@@ -330,28 +379,36 @@ agregarRecurso.addEventListener('click', () => {
       }  
     });
     
-  }  
+  }
+
+  Calcular(); 
 });
 
 //eliminacion de una fila correspondiente a un recurso
-eliminarRecurso.addEventListener('click', () => {  
-  if (recursos.focusActual != null) {
-    rolesRecursos.forEach(element => {
-      element.forEach((elemento, index) => {
-        if(elemento.rol == recursos.focusActual){
-          element.splice(index,1);
-        }
-      }); 
-    });
-    recursos.eliminarFila(recursos.focusActual);
-  }
-  else
-    alert("selecciona un rol de recurso primero");  
+eliminarRecurso.addEventListener('click', () => {
+  if (window.confirm("¿Esta seguro de que desea eliminar el recurso?")) {
+    if (recursos.focusActual != null) {
+      rolesRecursos.forEach(element => {
+        element.forEach((elemento, index) => {
+          if(elemento.rol == recursos.focusActual){
+            element.splice(index,1);
+          }
+        }); 
+      });
+      recursos.eliminarFila(recursos.focusActual);
+    }
+    else
+      alert("selecciona un rol de recurso primero");  
+    
+      //Calculo de valores tras eliminacion de recurso
+    Calcular();    
+  }  
 });
 
 
+//Proceso de guardado en la base de datos mediante API
 guardarPresupuesto.addEventListener('click', () => {
-  //console.log(mesesContemplados);
+  console.log(mesesContemplados);
   //console.log(conceptosIngresos);
   //console.log(conceptosCostos);
   //console.log(conceptosGastos);
@@ -361,15 +418,19 @@ guardarPresupuesto.addEventListener('click', () => {
 
 //Funcionalidad de calculo de valores en cada tabla (TEMPORAL)
 eliminarPresupuesto.addEventListener('click', () => { 
-
-  //console.log("Realizando operaciones");
-  recursos.calcularValorRecursos();
-  recursos.calcularValorCostos();
-  recursos.calcularValorResumen();
-
-  gastos.calcularValor();
-  costos.calcularValor();
-  ingresos.calcularValor();
-  estado.calcularValor();
+  //recursos.guardarValor(rolesRecursos);
+  //gastos.guardarValor(conceptosGastos);
+  //costos.guardarValor(conceptosCostos);
+  //ingresos.guardarValor(conceptosIngresos);
+  flujo.guardarValor(mesesContemplados, "abcdadw");
   
+});
+
+
+document.getElementById('botonLogout').addEventListener('click', async () =>{
+  localStorage.removeItem('ActiveUser');
+});
+
+document.getElementById('botonPresupuestos').addEventListener('click', async () =>{
+  window.location.href = "../html/main.html";
 });
